refactor: Moved VirtualScrollView to freya-components
