+++ components/src/lib.rs
@@ -5,6 +5,7 @@ mod scroll_view;
 mod slider;
 mod switch;
 mod theme;
+mod virtual_scroll_view;
 
 pub use button::*;
 pub use input::*;
@@ -13,3 +14,4 @@ pub use scroll_view::*;
 pub use slider::*;
 pub use switch::*;
 pub use theme::*;
+pub use virtual_scroll_view::*;
